// src/models/api/landing-page-api.ts
export interface ILandingPageRes {
    id: number;
    event_id: number;
    name: string;
    slug: string;
    title: string;
    subtitle: string;
    content: string;
    status: string;
    background_img: string;
    form_enable: number;
    form_title: string;
    form_description: string;
    form_input: string[];
    created_at: string;
    updated_at: string;
}